+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add remove directory button
 * fix some shiny bug
 * fix DT checkbox
+* shiny module : can now pass yml path
 
 # shinydrive 0.1.1